requirements.txt
asyncio
aiohttp
pyyaml
redis
